+++ src/index.js
@@ -1,5 +1,5 @@
 import onDispatchStoreEnhancer from './onDispatchStoreEnhancer'
 import withReducer from './withReducer'
-import { setContextType } from './contextTypes'
+import { setContextType } from './contextType'
 
 export { withReducer, onDispatchStoreEnhancer, setContextType }
